instead of having multiple accessor in one AggregationWithAccessor split it into
multiple independent AggregationWithAccessor

Performance seems unchanged
```
Before
test aggregation::agg_bench::bench::bench_aggregation_terms_many_json_mixed_type_with_sub_agg                            ... bench: 173,939,311 ns/iter (+/- 35,629,524)
test aggregation::agg_bench::bench::bench_aggregation_terms_many_json_mixed_type_with_sub_agg_multi                      ... bench: 227,312,130 ns/iter (+/- 34,251,990)
test aggregation::agg_bench::bench::bench_aggregation_terms_many_json_mixed_type_with_sub_agg_opt                        ... bench: 216,645,922 ns/iter (+/- 72,184,100)
test aggregation::agg_bench::bench::bench_aggregation_terms_many_json_mixed_type_with_sub_agg_sparse                     ... bench:  43,224,161 ns/iter (+/- 11,300,519)

After
test aggregation::agg_bench::bench::bench_aggregation_terms_many_json_mixed_type_with_sub_agg                            ... bench: 171,519,343 ns/iter (+/- 76,513,404)
test aggregation::agg_bench::bench::bench_aggregation_terms_many_json_mixed_type_with_sub_agg_multi                      ... bench: 228,347,835 ns/iter (+/- 12,738,742)
test aggregation::agg_bench::bench::bench_aggregation_terms_many_json_mixed_type_with_sub_agg_opt                        ... bench: 218,119,685 ns/iter (+/- 72,979,994)
test aggregation::agg_bench::bench::bench_aggregation_terms_many_json_mixed_type_with_sub_agg_sparse                     ... bench:  41,439,166 ns/iter (+/- 10,048,723)


```